+++ sdk/storage/storage-blob/test/blobclient.spec.ts
@@ -157,7 +157,7 @@ describe("BlobClient", () => {
 
     const result2 = (await containerClient
       .listBlobsFlat({
-        include: ["snapshots"]
+        includeSnapshots: true
       })
       .byPage()
       .next()).value;
@@ -175,7 +175,7 @@ describe("BlobClient", () => {
 
     const result3 = (await containerClient
       .listBlobsFlat({
-        include: ["snapshots"]
+        includeSnapshots: true
       })
       .byPage()
       .next()).value;
@@ -214,7 +214,7 @@ describe("BlobClient", () => {
 
     const result = (await containerClient
       .listBlobsFlat({
-        include: ["deleted"]
+        includeDeleted: true
       })
       .byPage()
       .next()).value;
@@ -225,7 +225,7 @@ describe("BlobClient", () => {
 
     const result2 = (await containerClient
       .listBlobsFlat({
-        include: ["deleted"]
+        includeDeleted: true
       })
       .byPage()
       .next()).value;